Remove App.tsx, add new fonts and svg assets
This commit is about file organization changes. App.tsx was removed since its rendering the whole components which is not efficient for our case, we prefer the Lazy Loading for each component separately. The fonts were added to support different styles in the design. The SVG assets - pizza, cart, plus icons were added to the public directory to be accessible for usage in the UI.